implementa service

+++ src/main/java/com/rlti/autoescola/servico/application/api/ServicoRequest.java
@@ -3,8 +3,10 @@ package com.rlti.autoescola.servico.application.api;
 import com.rlti.autoescola.servico.domain.Categoria;
 import lombok.Value;
 
+import java.math.BigDecimal;
+
 @Value
 public class ServicoRequest {
     Categoria categoria;
-    double valorServico;
+    BigDecimal valorServico;
 }
